+++ test/test_alloc/test_alloc_1.c
@@ -4,8 +4,8 @@
 
 static void proc2()
 {
-	size_t len1 = rand() % (SIZE) + 1;
-	size_t len2 = rand() % (SIZE) + 1;
+	size_t len1 = (size_t)rand() % (SIZE) + 1;
+	size_t len2 = (size_t)rand() % (SIZE) + 1;
 	void * buf1 = malloc(len1);                  ASSERT(buf1);
 	void * buf2 = malloc(len2);                  ASSERT(buf2);
 	memset(buf1, 0xFF, len1);
